pyproject.toml
[project]
name = "trackmania_tracker"
version = "3.2.2"
description = "The basic blog app built in the Flask tutorial."
dependencies = [
    "flask",
    "requests",
    "pyyaml",
    "pytz"
]
[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
